requirements.txt
av==8.0.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
cmake==3.22.1
cycler==0.11.0
decorator==4.4.2
dill==0.3.4
dlib==19.22.1
dominate==2.6.0
easydict==1.9
einops==0.3.2
ffmpeg==1.4
ffmpeg-python==0.2.0
flatbuffers==2.0
grpcio==1.43.0
imageio==2.13.3
imageio-ffmpeg==0.4.5
importlib-metadata==4.9.0
libclang==12.0.0
matplotlib==3.5.1
moviepy==1.0.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
proglog==0.1.9
protobuf==3.19.1
pyparsing==3.0.6
PySimpleGUI==4.55.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
